chore: refactor to one global vuestic vite plugin
