+++ packages/compiler/Readme.md
@@ -1,0 +1,31 @@
+# Vuestic Plugin
+
+Combines compilation-based utilities designed to improve development experience without impact on runtime
+
+## Vuestic Devtools
+
+Vuestic devtools designed for intuitive visual control over application with Vuestic components
+
+### Installation
+
+Install package
+
+Add `@vuestic/compiler/devtools/vite` plugin to `vite.config.ts`
+
+### Plans
+
+- [x] Edit component props
+- [x] Edit component slot content
+  - [] Add new slots
+- [] Add new components
+- [] Control layout
+- [] Add event listeners
+
+### Structure
+
+- Client - UI build with Vue and Vuestic
+- Vite - Vite plugin
+ - Compiler - compiler adds ids to components
+ - Server - can modify component source by id
+
+Later can be added plugin for Webpack with unplugin
