+++ packages/compiler/devtools/Readme.md
@@ -1,27 +1,5 @@
 # Vuestic Devtools
 
-Vuestic devtools designed for intuitive visual control over application with Vuestic components
-
-### Installation
-
-Install package
-
-Add `@vuestic/compiler/devtools/vite` plugin to `vite.config.ts`
-
-### Plans
-
-- [x] Edit component props
-- [x] Edit component slot content
-  - [] Add new slots
-- [] Add new components
-- [] Control layout
-- [] Add event listeners
-
-### Structure
-
-- Client - UI build with Vue and Vuestic
-- Vite - Vite plugin
- - Compiler - compiler adds ids to components
- - Server - can modify component source by id
-
-Later can be added plugin for Webpack with unplugin
+Client - UI for devtools, injected to user code in DEV mode only providing utilities to work visually with Vuestic components
+Server - used to update user's file trough client
+Plugin - adds vuestic devtools vue plugin to dev bundle and run server
