+++ packages/compiler/devtools/index.ts
@@ -1,0 +1,1 @@
+export { devtools, PluginOptions } from './plugin/plugin'
