+++ packages/compiler/devtools/plugin/add-vue-plugin.ts
