--- packages/compiler/devtools/shared/isHTMLElement.ts
@@ -1,3 +1,0 @@
-export const isHTMLElement = (value: unknown): value is HTMLElement => {
-  return value instanceof HTMLElement
-}
