--- packages/compiler/devtools/vite/index.ts
@@ -1,1 +1,0 @@
-export { vuesticDevtools } from './plugin';
